Move the resource type to the metric subsystem name (instead of a label)
This commit removes the `resource_type` label and moves it to the metric name.

Metric's names now use the following pattern:
1. `namespace` is a constant prefix (`stackdriver`)
2. `subsystem` is the normalized monitored resource type (ie `gce_instance`)
3. `name` is the normalized metric type (ie `compute_googleapis_com_instance_cpu_usage_time`)

If we don't set the resource_type as part of the metric name we get a
`has label dimensions inconsistent with previously collected metrics in the same metric family` error.
This is because the exporter attaches the monitored resource labels to the exported metric, and each
monitored resource type has a different set of labels. In prometheus a metric name should always expose
the same label names (with different values).

If we don't attach the monitored resource label to the metric, we get a
`was collected before with the same name and label values` error. This is because without the resource
labels we don't uniquelly identify the resource, so we might expose the same metric type with different
values.