Track delta metrics between scrapes (#168)
This PR introduces support for aggregating DELTA metrics in-memory. This implementation is somewhat based on https://github.com/prometheus-community/stackdriver_exporter/pull/167 but with extra functionality which is intended to solve the problem identified here, https://github.com/prometheus-community/stackdriver_exporter/pull/167#issuecomment-1217127124.

I chose to inject interfaces for the delta and distribution stores at the Collector level to try to keep library friendliness intact, https://github.com/prometheus-community/stackdriver_exporter/pull/157

I did as much as I could to explain the solution and the tradeoffs with aggregating deltas in this manner in the README.md. The bulk of the changes are in `monitoring_metrics.go` where I tried to obfuscate as much of the delta logic as possible.

Should hopefully resolve #116 and #25 

Signed-off-by: Kyle Eckhart <kgeckhart@users.noreply.github.com>