cache descriptors to reduce API calls
See issue #217

Currently the prefix -> descriptor mapping happens on every /metrics
call. In our setup we have a lot of exporters running with specific
prefixes that usually map to only a single/handful of metrics. This
makes the ratio of ListDescriptor to ListTimeseries calls often close to
50/50 (43% of total request are ListDescriptor in a random sampling).

In our case we're only interested in scraping the metrics produced by
google services. Those services very rarely change their metrics so
fetching the descriptors every time is just wasted work.
While its only a few calls per exporter it adds up to millions of calls
a month (2 (prometheuses in HA) * 1 (scrape per minute) * 60 * 24 * 31
(hours per month) * 20 prefixes = 1'728'000), across many google
projects.

This commit introduces a simple prefix -> []descriptor cache. If the
prefix isn't in the cache the original page based lookup is performed
and at the end of it the cache is updated. If it does exist in the cache
the cached value is used.
No attempt is made to prevent two routines from updating the cache
concurrently. If two concurrent requests see an expired cache entry
they'll both do the ListDescriptor call and update the cache. Doing this
keeps the cache simple and doesn't block one request on the other.
This situation will also be rare in practice, as there is usually only
one or two prometheuses scraping.

The cache itself comes in 3 flavors:

- noop, does nothing
- descriptorCache, caches all descriptors
- googleDescriptorCache, caches only descriptors of google metrics

The cache is disabled by default and can be enabled with the new
`--monitoring.descriptor-cache-ttl` CLI flag, if the TTL is set to a
value >0s the cache is enabled. It defaults to using the google only
cache, to control that the
`--no-monitoring.descriptor-cache-only-google` is introduced

Signed-off-by: bas smit <bsmit@bol.com>