feat : Add Support for Non-Default Google Cloud Universes via a flag #429
This commit adds a new flag, `google.universe- domain`, which allows users to specify the Google Cloud universe to use. This is useful for  users who are not using the default `googleapis.com` universe.

The new flag is used when creating the monitoring service, so that the service is configured to use the specified universe.

Signed-off-by: Mathieu Goulin <matgou1@msn.com>